pyproject.toml
[project]
name = "finbot"
version = "0.1.0"
description = "Telegram bot for stock analysis and forecasting"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.0",
    "yfinance>=0.2.48",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "scikit-learn>=1.5.2",
    "prophet>=1.1.6",
    "torch>=2.5.1",
    "matplotlib>=3.9.2",
    "scipy>=1.14.1",
    "python-dotenv>=1.0.1",
    "scipy-stubs==1.16.3.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.4",
    "pytest>=8.3.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["X", "X_*"]
